pyproject.toml
[tool.poetry]
name = "billing-service"
version = "0.1.0"
description = "Billing system for online movies service"
authors = ["team1 <team1@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "billing-service"}]

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
src_paths = ["src/app"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
pydantic = "^2.1.1"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
gunicorn = "^21.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
orjson = "^3.9.2"
uvicorn = "^0.23.1"
redis = "^4.6.0"
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
bandit = "^1.7.5"
mypy = "^1.4.1"
flake8-bandit = "^4.1.1"
flake8-print = "^5.0.0"
flake8-broken-line = "^1.0.0"
flake8-bugbear = "^23.7.10"
flake8-commas = "^2.1.0"
pre-commit = "^3.3.3"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
black = "^23.7.0"
isort = "^5.12.0"
types-redis = "^4.6.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
